pyproject.toml
[tool.poetry]
name = "sensor-dashboard"
version = "0.1.0"
description = ""
authors = ["Aditya W <07311940000001@mahasiswa.integra.its.ac.id>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.3"
streamlit = "^1.3.0"
pyserial = "^3.5"
plotly = "^5.6.0"
ipykernel = "^6.10.0"
jupyter = "^1.0.0"
pyqtgraph = "^0.12.4"
PyQt5 = "^5.15.6"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
